requirements.txt
asgiref==3.11.0
Django==6.0
pdf2image==1.17.0
pillow==12.1.0
PyMuPDF==1.26.7
sqlparse==0.5.5
tzdata==2025.3
